Clean accounts dirs earlier
