+++ validator/Cargo.toml
@@ -29,9 +29,10 @@ solana-download-utils = { path = "../download-utils", version = "1.6.0" }
 solana-faucet = { path = "../faucet", version = "1.6.0" }
 solana-ledger = { path = "../ledger", version = "1.6.0" }
 solana-logger = { path = "../logger", version = "1.6.0" }
-solana-perf = { path = "../perf", version = "1.6.0" }
+solana-measure = { path = "../measure", version = "1.6.0" }
 solana-metrics = { path = "../metrics", version = "1.6.0" }
 solana-net-utils = { path = "../net-utils", version = "1.6.0" }
+solana-perf = { path = "../perf", version = "1.6.0" }
 solana-runtime = { path = "../runtime", version = "1.6.0" }
 solana-sdk = { path = "../sdk", version = "1.6.0" }
 solana-version = { path = "../version", version = "1.6.0" }
